Create and implemented w6.cpp
w6.cpp contains the main() which hosts a lambda expression that will be passed to the display() funtion. This lambda expression converts numeric grade to a letter grade. The program will read from a file and output to the console the student number, numeric grade and the converted letter grade.